==> Lombardia/LarioLecco/2-GeoReferenceDataCollection.py <==
import acqua.geoCode as gc
import acqua.aqueduct as aq
gestore = "LarioLecco"
aq.setEnv('Lombardia//'+gestore)
locationListFile = 'Metadata/LocationListReviewed.csv'
geoReferencedLocationsFile = 'Metadata/GeoReferencedLocationsList.csv'
reteAquedottiFile = '../ReteAcquedotti/rete_acquedotti_lombardia(escl_Milano).geojson'
mapFile = gestore+'.html'
gc.createGeoReferencedLocationsList(locationListFile,geoReferencedLocationsFile)
gc.findGeoName(geoReferencedLocationsFile)
gc.findCoordinates(geoReferencedLocationsFile)
gc.createMapOnGeoReferencedDataCollection(geoReferencedLocationsFile,reteAquedottiFile,mapFile)
